requirements/dev.txt
-r base.txt
django-shortcuts==1.5
